// src/types/navigation.ts
export type ViewType = 
  | 'published' 
  | 'all' 
  | 'pab-dashboard' 
  | 'management-dashboard' 
  | 'program-dashboard'
  | 'risk-dashboard'
  | 'quality-dashboard'
  | 'scope-dashboard'
  | 'safety-dashboard'
  | 'environment-dashboard'
  | 'people-dashboard'
  | 'cost-dashboard'
  | 'operational-dashboard';